refactor(saml): modularize SamlApplication
